+++ .vscode-test.mjs
@@ -5,6 +5,7 @@ export default defineConfig({
   workspaceFolder: './src/test/intellisense',
   version: '1.98.2',
   mocha: {
+    // timeout: 20_000_000_000
     timeout: 20_000
   }
 })