+++ src/test/intellisense/helper.ts
@@ -1,4 +1,7 @@
+import * as path from 'path'
 import * as vscode from 'vscode'
+
+import { compile } from '../../compiler/compiler'
 
 export const __dirnameProxy = __dirname.replace(/([\\\/]+)out([\\\/]+)/, '$1src$2')
 
@@ -20,4 +23,20 @@ export async function waitForSelectionChange({
       resolve(false) // Разрешаем промис с false
     }, timeoutInMilliseconds)
   })
+}
+export async function buildWithTreeShaking(testCases: string) {
+  const config = {
+    entry: path.join(__dirnameProxy, 'tree-shaking', testCases, 'entry.js'),
+    projectRoot: path.join(__dirnameProxy, 'tree-shaking', testCases)
+  }
+
+  const sn = await compile(config.entry, {
+    projectRoot: config.projectRoot,
+    modules: [],
+    treeShaking: { searchPattern: /double/ },
+    skipAttachGlobalScript: true
+  })
+  const bundle = sn.toStringWithSourceMap().code
+
+  return bundle.replace(/(\n|\r\n)/gm, '\n')
 }