+++ src/test/intellisense/tree-shaking/direct-inheritance-reverse-import/double.js
@@ -1,0 +1,5 @@
+import './constants.js'
+
+function double(a) {
+  return a * 2
+}