+++ src/test/intellisense/tree-shaking/direct-inheritance-reverse-import/entry.js
@@ -1,0 +1,4 @@
+import './sum.js'
+import './double.js'
+
+double(2)