+++ src/test/intellisense/tree-shaking/direct-inheritance/entry.js
@@ -1,0 +1,4 @@
+import './double.js'
+import './sum.js'
+
+double(2)