+++ src/test/intellisense/tree-shaking/indirect-inheritance-reverse-import/constants.js
@@ -1,0 +1,1 @@
+const CONSTANT = 0