+++ src/test/intellisense/tree-shaking/indirect-inheritance-reverse-import/entry.js
@@ -1,0 +1,3 @@
+import './dep.js'
+
+double(2)