+++ src/test/intellisense/tree-shaking/indirect-inheritance/dep.js
@@ -1,0 +1,2 @@
+import './double.js'
+import './sum.js'