+++ src/test/intellisense/tree-shaking/indirect-inheritance/sum.js
@@ -1,0 +1,5 @@
+/** @module math */
+
+function sum(a, b) {
+  return a + b
+}